dma-heap: align fd_flags and heap_flags with dma_heap_allocation_data
According to commit 310ec03841a3 ("dma-buf: align fd_flags and heap_flags with dma_heap_allocation_data").

+++ include/linux/dma-heap.h
@@ -76,6 +76,6 @@ struct dma_heap *dma_heap_find(const char *name);
  * This is for internal dma-buf allocations only.
  */
 struct dma_buf *dma_heap_buffer_alloc(struct dma_heap *heap, size_t len,
-				      unsigned int fd_flags,
-				      unsigned int heap_flags);
+				      u32 fd_flags,
+				      u64 heap_flags);
 #endif /* _DMA_HEAPS_H */
